update доки и конфиги сервера авторизации
