+++ .env
@@ -1,3 +1,6 @@
+# common
+JWT_SECRET_WORD=Vityalol
+
 # redis
 REDIS_HOST=apple_pie_redis
 REDIS_PORT=6379
