+++ auth_server/.dockerignore
@@ -1,4 +1,5 @@
 Dockerfile
 .dockerignore
 node_modules
-API_DOCS.md
+API_DOC.md
+README.md
