+++ auth_server/.env
@@ -1,0 +1,1 @@
+SESSION_EXPIRES_IN=259200
