+++ auth_server/README.md
@@ -1,5 +1,9 @@
 # Authotization server for @Apple_pie project
 
+## Params
+
+ * **session expires in**: 30 days (`.env` config)
+
 ## Tools
 
  * `docker exec -it apple_pie_auth_server sh` enter container
@@ -7,4 +11,5 @@
 
 ## Docs
 
- * [RestAPI doc](./API_DOCS.md).    
+ * [RestAPI](./API_DOC.md)    
+ * [DB schema (models)](../redis/SCHEMA_DOC.md)
