--- auth_server/src/config.js
@@ -1,1 +1,0 @@
-export const SESSION_EXPIRES_IN = 259200; // 360 * 24 * 30
