+++ nginx/.dockerignore
@@ -1,4 +1,3 @@
 Dockerfile
 .dockerignore
 README.md
-.env
