+++ redis/.dockerignore
@@ -1,2 +1,4 @@
 Dockerfile
 .dockerignore
+README.md
+SCHEMA_DOC.md
