+++ redis/README.md
@@ -13,3 +13,7 @@
  * `docker exec -it apple_pie_redis sh` enter container
  * `docker logs -f apple_pie_redis` watch logs
  * `docker exec -it apple_pie_redis redis-benchmark -q -n 100000 -c 50 -P 12` performance banchmark (Выполнит 100 000 запросов от 50 клиентов по 12 команд одновременно)
+
+## Docs
+
+ * [DB schema (models)](./SCHEMA_DOC.md)
