+++ root_server/.dockerignore
@@ -1,4 +1,4 @@
 Dockerfile
 .dockerignore
 node_modules
-API_DOCS.md
+API_DOC.md
