+++ root_server/src/index.ts
@@ -1,4 +1,4 @@
-import 'module-alias/register'; // для алиасов путей
+import 'module-alias/register'; // для алиасов путей TS
 import { config } from 'dotenv'; config();
 
 import app from './app';
